% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Bucket}
\alias{Bucket}
\title{Cloud Storage JSON API Objects 
Lets you store and retrieve potentially-large, immutable data objects.}
\usage{
Bucket(Bucket.owner = NULL, Bucket.website = NULL, acl = NULL,
  defaultObjectAcl = NULL, id = NULL, location = NULL, owner = NULL,
  projectId = NULL, selfLink = NULL, timeCreated = NULL, website = NULL)
}
\arguments{
\item{Bucket.owner}{The \link{Bucket.owner} object or list of objects}

\item{Bucket.website}{The \link{Bucket.website} object or list of objects}

\item{acl}{Access controls on the bucket}

\item{defaultObjectAcl}{Default access controls to apply to new objects when no ACL is provided}

\item{id}{The name of the bucket}

\item{location}{The location of the bucket}

\item{owner}{The owner of the bucket}

\item{projectId}{The project the bucket belongs to}

\item{selfLink}{The URI of this bucket}

\item{timeCreated}{Creation time of the bucket in RFC 3339 format}

\item{website}{The bucket's website configuration}
}
\value{
Bucket object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:17:26
filename: /Users/mark/dev/R/autoGoogleAPI/googlestoragev1beta1.auto/R/storage_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Bucket Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A bucket.
}
\seealso{
Other Bucket functions: \code{\link{Bucket.owner}},
  \code{\link{Bucket.website}},
  \code{\link{buckets.insert}},
  \code{\link{buckets.patch}}, \code{\link{buckets.update}}
}
