% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{buckets.patch}
\alias{buckets.patch}
\title{Updates a bucket. This method supports patch semantics.}
\usage{
buckets.patch(Bucket, bucket, projection = NULL)
}
\arguments{
\item{Bucket}{The \link{Bucket} object to pass to this method}

\item{bucket}{Name of a bucket}

\item{projection}{Set of properties to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/devstorage.full_control, https://www.googleapis.com/auth/devstorage.read_write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}

Other Bucket functions: \code{\link{Bucket.owner}},
  \code{\link{Bucket.website}}, \code{\link{Bucket}},
  \code{\link{buckets.insert}},
  \code{\link{buckets.update}}
}

