% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Object.media}
\alias{Object.media}
\title{Object.media Object}
\usage{
Object.media(algorithm = NULL, contentType = NULL, data = NULL,
  hash = NULL, length = NULL, link = NULL, timeCreated = NULL)
}
\arguments{
\item{algorithm}{Hash algorithm used}

\item{contentType}{Content-Type of the object data}

\item{data}{URL-safe Base64-encoded data}

\item{hash}{Hash of the data}

\item{length}{Content-Length of the data in bytes}

\item{link}{Media download link}

\item{timeCreated}{Creation time of the data in RFC 3339 format}
}
\value{
Object.media object
}
\description{
Object.media Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Object media data. Provided on your behalf when uploading raw media or multipart/related with an auxiliary media part.
}
\seealso{
Other Object functions: \code{\link{Object.metadata}},
  \code{\link{Object.owner}}, \code{\link{Object}},
  \code{\link{objects.insert}},
  \code{\link{objects.patch}}, \code{\link{objects.update}}
}
