% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{objects.list}
\alias{objects.list}
\title{Retrieves a list of objects matching the criteria.}
\usage{
objects.list(bucket, delimiter = NULL, max.results = NULL,
  pageToken = NULL, prefix = NULL, projection = NULL)
}
\arguments{
\item{bucket}{Name of the bucket in which to look for objects}

\item{delimiter}{Returns results in a directory-like mode}

\item{max.results}{Maximum number of items plus prefixes to return}

\item{pageToken}{A previously-returned page token representing part of the larger set of results to view}

\item{prefix}{Filter results to objects whose names begin with this prefix}

\item{projection}{Set of properties to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_only
\item https://www.googleapis.com/auth/devstorage.read_write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/devstorage.full_control, https://www.googleapis.com/auth/devstorage.read_only, https://www.googleapis.com/auth/devstorage.read_write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}
}
