% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Objects}
\alias{Objects}
\title{Objects Object}
\usage{
Objects(items = NULL, nextPageToken = NULL, prefixes = NULL)
}
\arguments{
\item{items}{The list of items}

\item{nextPageToken}{The continuation token, used to page through large result sets}

\item{prefixes}{The list of prefixes of objects matching-but-not-listed up to and including the requested delimiter}
}
\value{
Objects object
}
\description{
Objects Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of objects.
}
